+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added emoji picker to post editor.
 - Added "Subscriber details" page.
+- Allow to add subscribers and extend existing subscriptions.
 
 ## [2.14.0] - 2024-03-30